64-bit ints

+++ examples/ints.c
@@ -1,3 +1,4 @@
+#include <stdint.h>
 #include <stdio.h>
 
 #include "hashtable.h"
@@ -10,11 +11,11 @@ int main() {
     hashtable_set_int(my_hashtable, "Apples", 2000);
     hashtable_set_int(my_hashtable, "Oranges", 1000);
 
-    int apples, oranges;
+    int64_t apples, oranges;
     hashtable_get_int(my_hashtable, &apples, "Apples");
     hashtable_get_int(my_hashtable, &oranges, "Oranges");
 
-    printf("Apples: %d\nOranges: %d\n", apples, oranges);
+    printf("Apples: %ld\nOranges: %ld\n", apples, oranges);
 
     hashtable_free(my_hashtable);
